Change Storybook import
