+++ index.app.js
@@ -1,0 +1,7 @@
+import { AppRegistry } from 'react-native';
+
+import { name as appName } from './app.json';
+
+import App from './src/App';
+
+AppRegistry.registerComponent(appName, () => App);
