+++ index.js
@@ -1,9 +1,2 @@
-import { AppRegistry } from 'react-native';
-
-import App from './src/App';
-import { name as appName } from './app.json';
-
-AppRegistry.registerComponent(appName, () => App);
-
-// import StorybookNavigator from './storybook';
-// export default StorybookNavigator;
+import './index.app';
+// import './index.storybook';
