+++ storybook/stories/welcome.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react-native';
+
+import Welcome from '../components/Welcome';
+
+storiesOf('Welcome', module).add('to Storybook', () => <Welcome />);
